Refactor defining above and below shape keys in merge operator
Not only was the code overly complex, but it was wrong in number of ways: above and below tags were assigned incorrectly (should've been vice versa), and in single merge operator it had unnecessary iteration to get all above or below shape keys to set their values to 0, when in fact it should've set value to 0 for all shape keys.

Changing that means that single merge operator now gives correct results in almost all simple cases.