Don't set filtered shape key values to 1 when merging
I think this isn't necessary at all. Setting all values to 1 and creating shape key from that mix creates a variant that never existed before, while keeping current values results in a similar looking shape key.

In fact, look is exactly the same, but the problem is value, min, and max need to be adjusted to get the exact same result, especially for animation, but I'll leave that for future, because its quite a large change and will require recreating entire animation.